+++ CHANGELOG.md
@@ -6,6 +6,7 @@ HEAD
 
 * Fixed an access violation in `DynamicJsonBuffer` when memory allocation fails (issue #433)
 * Added operators `==` and `!=` for two `JsonVariant`s (issue #436)
+* Fixed `JsonVariant::operator[const FlashStringHelper*]` (issue #441)
 
 v5.8.2
 ------