+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ArduinoJson: change log
 =======================
 
-HEAD
-----
+v5.8.3
+------
 
 * Fixed an access violation in `DynamicJsonBuffer` when memory allocation fails (issue #433)
 * Added operators `==` and `!=` for two `JsonVariant`s (issue #436)